added Suggests: wasta-offline-setup to indicate companion app.
wasta-offline-setup helps to create and clean the USB cache used
by wasta-offline.

wasta-offline-setup is not needed on every machine and has extra
dependencies that are not often pre-installed, so it was broken
out as a separate app.

wasta-offline is a recommendation in wasta-offline-setup.  A
reverse recommendation would be too strong and typically autoinstalled
so instead wasta-offline suggests wasta-offline-setup in order to
alert the installer about the companion app.