.pkla is obsolete in noble, hint with new javascript .rules format
It was hard to debug this. I tried lots of web suggestions,
but in the end, I killed polkitd and ran it interactively.
(I don't know if the other stuff I did actually came into play...)

G_MESSAGES_DEBUG=all /usr/lib/policykit-1/polkitd

Getting the permissions right seems important too.
Based on the directory, it should be 660 root:polkitd.

The other annoying thing is that even though /var/log/auth.log
shows polkit re-reading all the rules as soon as it changes,
it didn't come into effect until after systemctl restart polkit!

A third annoying thing is that it seems to be necessary to enclose
in round brackets the == clauses, so I did that already
inside the if() clause to encourage keeping it in case an
&& subject.isInGroup("sudo")
or some other condition is added.