2.4.0: Don't share repo folders between Mint and Ubuntu
This patch must go hand-in-hand with wasta-offline-setup.
Surprisingly, we never actually created a MINT_SERIES folder,
so the only time that would have been made available
was if someone hand-created that.

If wasta-offline is used at all for Linux Mint,
the likelihood of someone sharing the offline cache between
Ubuntu and Mint machines is very high,
since Wasta is based on Ubuntu.

Well, that is just a recipe for disaster.
In my brief use of this with Mint 22.1 xia,
I saw the lsb_release info change to Ubuntu settings,
and that can't be a good thing.

Plus, with Wasta's use of cinnamon packages,
it could potentially be a real problem if Ubuntu's folder
should somehow get newer versions than what Mint is using.

So, just bite the USB-disk-space-usage bullet
and keep entirely separate folders for Mint and Ubuntu.